build: Support global constructors
This commit implements crti.S and crtn.S. These functions are needed to
support initialisation and termination routines such as for C++ binaries
or for C compiled with the GCC constructor attribute.

Tested by adding a constructor function to kernel.c which initializes
and writes to the terminal, and seeing that it was called.