feat: Setup global descriptor table
This commit adds C and assembly routines which allocate memory for a GDT
and initialize it. The memory for the GDT is allocated on the kernel
stack by a C routine, it might be nice later to bake it into boot.S like
xv6 does.

Kernel does not crash, therefore we have cause to be vaguely optimistic.